refactor: address additional Copilot feedback
- Extract cache invalidation logic into invalidate_schema_cache helper
- Create shared SchemaHelpers module to eliminate code duplication
- Consolidate additionalProperties logic across Tool and OutputSchemaBuilder
- Fix rubocop style issues (trailing whitespace, empty lines)

Addresses all remaining Copilot review suggestions from PR #142
